auth done

+++ types/authorization.go
@@ -7,5 +7,5 @@ type ParseTokenWorkstationResponse struct {
 
 type SessionInfo struct {
 	RefreshToken string `json:"refresh_token" db:"refresh_token"`
-	ExpiresAt    int    `json:"expires_at" db:"expires_at"`
+	ExpiresAt    int64  `json:"expires_at" db:"expires_at"`
 }
